# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "FieldErrorBody": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "field": {
            "title": "Field",
            "type": "string"
          }
        },
        "required": [
          "field",
          "code"
        ],
        "title": "FieldErrorBody",
        "type": "object"
      },
...
      "ProblemDetails": {
        "description": "OpenAPI description of the error body (the wire shape is built by ``problem_response``).",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FieldErrorBody"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errors"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code",
          "detail"
        ],
        "title": "ProblemDetails",
        "type": "object"
      },
...
      "QueuePageOut": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/QueueItemOut"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "QueuePageOut",
        "type": "object"
      },
...
      "ReportReceipt": {
        "additionalProperties": false,
        "description": "Shown once. The client must display the code and never put it in a URL or storage.",
        "properties": {
          "attachments": {
            "items": {
              "$ref": "#/components/schemas/AttachmentOutcome"
            },
            "title": "Attachments",
            "type": "array"
          },
          "contact_saved": {
            "title": "Contact Saved",
            "type": "boolean"
          },
          "next_steps": {
            "items": {
              "enum": [
                "save_tracking_code",
                "check_status_later",
                "see_escalation_guidance"
              ],
              "type": "string"
            },
            "title": "Next Steps",
            "type": "array"
          },
          "published": {
            "const": false,
            "title": "Published",
            "type": "boolean"
          },
          "status": {
            "const": "received",
            "title": "Status",
            "type": "string"
          },
          "tracking_code": {
            "title": "Tracking Code",
            "type": "string"
          }
        },
        "required": [
          "tracking_code",
          "status",
          "published",
          "contact_saved",
          "attachments",
          "next_steps"
        ],
        "title": "ReportReceipt",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "ShaidaGo platform API",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/reports": {
      "post": {
        "operationId": "reports_submit",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "attachments": {
                    "items": {
                      "format": "binary",
                      "type": "string"
                    },
                    "maxItems": 3,
                    "type": "array"
                  },
                  "concern_category": {
                    "enum": [
                      "no_visible_work",
                      "incomplete_work",
                      "unsafe_construction",
                      "suspected_incorrect_status",
                      "access_barrier",
                      "other_concern"
                    ],
                    "type": "string"
                  },
                  "contact_channel": {
                    "enum": [
                      "email",
                      "phone",
                      "messaging_app"
                    ],
                    "type": "string"
                  },
                  "contact_value": {
                    "maxLength": 200,
                    "type": "string"
                  },
                  "description": {
                    "maxLength": 8000,
                    "minLength": 10,
                    "type": "string"
                  },
                  "project_slug": {
                    "maxLength": 120,
                    "type": "string"
                  },
                  "reporter_handle": {
                    "maxLength": 200,
                    "type": "string"
                  },
                  "reporter_passphrase": {
                    "maxLength": 200,
                    "type": "string"
                  }
                },
                "required": [
                  "project_slug",
                  "concern_category",
                  "description"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportReceipt"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Missing or malformed Idempotency-Key."
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Reporter handle credentials not accepted."
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Key reused with a different request."
          },
          "413": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Content Too Large"
          },
          "415": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unsupported Media Type"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Submit Report",
        "tags": [
          "reports"
        ]
      }
    },
...
    "/v1/reviewer/reports": {
      "get": {
        "operationId": "reviewer_reports_queue",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "enum": [
                      "received",
                      "needs_information",
                      "under_review",
                      "verified_for_public_update",
                      "referred",
                      "closed"
                    ],
                    "type": "string"
                  },
                  "maxItems": 6,
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "risk_level",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "standard",
                    "elevated",
                    "high"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Risk Level"
            }
          },
          {
            "in": "query",
            "name": "concern_category",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "no_visible_work",
                    "incomplete_work",
                    "unsafe_construction",
                    "suspected_incorrect_status",
                    "access_barrier",
                    "other_concern"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Concern Category"
            }
          },
          {
            "in": "query",
            "name": "project",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 120,
                  "minLength": 1,
                  "pattern": "^[a-z0-9-]+$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 1000,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 512,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueuePageOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Invalid pagination cursor."
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Queue",
        "tags": [
          "reviewer"
        ]
      }
    },
    "/v1/reviewer/reports/{report_id}": {
      "get": {
        "operationId": "reviewer_reports_get",
        "parameters": [
          {
            "in": "path",
            "name": "report_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Report Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "include_contact",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Contact",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportDetailOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Invalid pagination cursor."
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Detail",
        "tags": [
          "reviewer"
        ]
      }
    },
...
  }
}